Fix disposition font problem
The disposition font can not be be extended on scrlttr classes, so we
now test to ensure we're in the scrartcl class before we attempt to
extend the disposition font. The scrbook and other classes would have to
be explicitly defined. I considered changing the packages I have into
document classes that are dependent on the KOMA-Script classes but the
problem there is that the style package would remain a package and not a
class. If I want to do this properly I would have to move some elements
into the bilingualagreement or longinvoice packages (which will become
the documentclasses) which couples styling too much with structure.

The bilingualagreement and longinvoice packages are primarily
structure-oriented, while asabina-style is very much styling oriented.
Moving `\addtokomafont` calls into the layout documents feels a bit like
an overreach for a file defining the layout.